refactor: corrige importações e refatora componentes
