+++ src/pages/Dashboard/index.tsx
@@ -1,5 +1,5 @@
-import { Footer } from '../../components/templates/Footer';
-import { NavBar } from '../../components/templates/NavBar';
+import { Footer } from '../../components/__templates__/Footer';
+import { NavBar } from '../../components/__templates__/NavBar';
 
 export const Dashboard = () => {
   return (
